eduforge/core/billing/apps.py
"""
Billing app configuration.
"""
from django.apps import AppConfig


class BillingConfig(AppConfig):
    default_auto_field = 'django.db.models.BigAutoField'
    name = 'core.billing'
    verbose_name = 'Billing'
