eduforge/requirements/dev.txt
# Development dependencies
-r base.txt
django-debug-toolbar==4.2.0
black==23.12.1
flake8==6.1.0
isort==5.13.2
pytest==7.4.3
pytest-django==4.7.0
faker==21.0.0
